Refactor: chat의 main부분 폴더로 구조화
