+++ src/components/chat/messageBox/CharacterSpeak.tsx
